+++ src/main/resources/META-INF/resources/templates/footer.xhtml
@@ -1,0 +1,18 @@
+<div class="layout-footer">
+    <div class="layout-footer-item1">
+        <span>v1.0.0-SNAPSHOT by </span>
+        <a href="https://www.primetek.com.tr">R3ing</a>
+    </div>
+
+    <div class="layout-footer-item1">
+        <a href="https://github.com/primefaces/primefaces" class="mr-3" title="GitHub">
+            <i class="pi pi-github"></i>
+        </a>
+        <a href="https://twitter.com/primefaces" class="mr-3" title="Twitter">
+            <i class="pi pi-twitter"></i>
+        </a>
+        <a href="https://discord.com/invite/gzKFYnpmCY" title="Discord">
+            <i class="pi pi-discord"></i>
+        </a>
+    </div>
+</div>
