+++ src/main/resources/META-INF/resources/templates/header.xhtml
@@ -1,0 +1,83 @@
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:h="http://xmlns.jcp.org/jsf/html"
+      xmlns:p="http://primefaces.org/ui">
+<h:body>
+
+    <div class="layout-topbar">
+        <a class="menu-button" tabindex="0">
+            <i class="pi pi-bars"></i>
+        </a>
+
+        <a href="${pageContext.request.contextPath}/system/home.xhtml" class="logo" title="Showcase Home">
+            <img id="j_idt27" src="${pageContext.request.contextPath}/assets/icons/cfe-icon.png" alt="logo"
+                 style="width: 4rem; height: auto; margin-left: 2rem;"/>
+        </a>
+
+        <form class="topbar-form">
+            <ul class="topbar-menu">
+                <li><a href="/system/home.xhtml">Customers</a></li>
+                <li class="topbar-submenu"><a href="/system/home.xhtml" tabindex="0">Shipppings</a></li>
+                <li class="topbar-submenu"><a href="/system/home.xhtml" tabindex="0">Resources</a></li>
+
+                <li class="topbar-submenu">
+                    <p:menubar styleClass="topbar-submenu">
+                        <p:submenu icon="pi pi-users">
+                            <p:menuitem icon="pi pi-user"
+                                    value="Hello #{userBean.getUsername().substring(0, userBean.getUsername().lastIndexOf('@'))}!"/>
+                            <p:divider/>
+
+<!--                            <p:menuitem  value="Change Password" icon="pi pi-key"  actionListener="#{userBean.showChangePasswordDialog}" update="changePasswordDialog"/>-->
+
+                            <p:menuitem icon="pi pi-power-off" value="Logout" url="/logout"/>
+                        </p:submenu>
+                    </p:menubar>
+                </li>
+
+            </ul>
+        </form>
+    </div>
+
+    <p:dialog header="Change Password"
+              id="changePasswordDialog"
+              widgetVar="changePasswordDialog"
+              modal="true"
+              resizable="false"
+              appendTo="@(body)"
+              rendered="#{userBean.showPasswordDialog}"
+              visible="#{userBean.showPasswordDialog}"
+              onShow="PF('changePasswordDialog').hide();">
+        <h:form id="changePasswordForm">
+            <p:panelGrid columns="2" styleClass="ui-noborder">
+                <p:outputLabel for="currentPassword" value="Current Password:"/>
+                <p:password id="currentPassword" value="#{userBean.currentPassword}"
+                            required="true"/>
+
+                <p:outputLabel for="newPassword" value="New Password:"/>
+                <p:password id="newPassword" value="#{userBean.newPassword}"
+                            required="true"/>
+
+                <p:outputLabel for="confirmPassword" value="Confirm Password:"/>
+                <p:password id="confirmPassword" value="#{userBean.confirmPassword}"
+                            required="true"/>
+            </p:panelGrid>
+
+            <p:separator/>
+
+            <p:commandButton value="Save"
+                             icon="pi pi-check"
+                             styleClass="ui-button-outlined"
+                             action="#{userBean.changePassword}"
+                             update="changePasswordForm"
+                             oncomplete="PF('changePasswordDialog').hide()"/>
+            <p:commandButton value="Cancel"
+                             onclick="PF('changePasswordDialog').hide()"
+                             type="button"
+                             icon="pi pi-sign-out"
+                             styleClass="ui-button-outlined ui-button-secondary"
+                             style="margin-left:10px;"/>
+        </h:form>
+    </p:dialog>
+
+</h:body>
+</html>
